+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Success toast message is not shown at the start of export. Refs UIOR-1089.
 * The `limit` query parameter is not explicitly set to some version history's requests. Refs UIOR-1101.
 * Keyboard shortcuts issues. Refs UIOR-1086.
+* A user can save duplicated order template name with only spaces before and after initial template name. Ref UIOR-1114
 
 ## [4.0.2](https://github.com/folio-org/ui-orders/tree/v4.0.2) (2023-03-17)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v4.0.1...v4.0.2)